GEO-INFER-BIO/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
biopython>=1.79
networkx>=2.6.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
geopandas>=0.9.0
shapely>=1.8.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
graphql-core>=3.1.0
strawberry-graphql>=0.96.0
geo-infer-space>=0.1.0
geo-infer-time>=0.1.0
geo-infer-ai>=0.1.0

# Development dependencies
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7b0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910
sphinx>=4.1.0
sphinx-rtd-theme>=0.5.0 